Menu and Shift (#27)
* Add notification icon and user dropdown menu to navbar

The update includes a notification bell icon addition and a user avatar dropdown menu enhancement in the navbar. The dropdown menu now incorporates options for password change, settings access, and logout. Existing navbar design principles were followed for the layout and styling of the new components.

* Add conditional renderings of SVG elements on navbar

Added logic to conditionally render different SVG icons based on user shift status on the navigation bar. The SVG icons will change depending on whether the user is "on_duty". This update enhances the UI by providing realtime visual cues to the users about their session status.

* Add time tracking function and shift-start/end interface

Added a `createTimeTracker` function in `User` model and a `Shift` livewire component to handle shift-start and -end. Also updated the front-end interface to enable a user to end his/her shift. The `TimeTracker` model has been slightly modified to align with these changes.

* Hide user-menu by default in nav component

Added "display:none" style to the user-menu div in the nav.blade.php file. This change hides the user-menu by default, improving the UI by showing the user-menu only when needed.

* Remove shift UI code redundancy, refactor to Livewire component

Removed duplicated shift-related UI code in the app's navigation components. Shift status/button UI is now refactored into a dedicated Livewire component: 'livewire.shift'. This enhances code reusability and makes UI updates more centralized and efficient.

* Add @alpinejs/ui dependency and update UI in app.js

Implemented the @alpinejs/ui plugin by adding it as a dependence in package.json and package-lock.json. The UI in app.js has been updated to use this plugin. This change is aimed at improving interaction and user experience, as well as enhancing efficiency in code execution.

* Replace HasMany relationship with BelongsToMany

This commit replaces the HasMany relationship with BelongsToMany in the User and Location models. This change is important to correctly establish the many-to-many relationship between these two entities.

* Refactor shift scheduling and add dark mode support

The changes introduced in this commit include a significant refactor of the shift scheduling feature in the Shift component. A dialog box for shift start time and location has been added to the UI and validation has been implemented on these inputs. Additionally, support for dark mode has been included in the application's views and components.

* Add user authentication for logout button test

The current commit introduces a test change in the Dashboard component. Previously the logout button test was not running on the authentication basis making the test faulty. Now, 'actingAs' Laravel function has been used to authenticate a user before running the test.

* Implement checkpoint validation to startShift function

This update adds a validation line before the starting of a shift in the startShift function. In addition, it modifies the conditions within the shift blade's logout button, now only responding when a location count is present for an authenticated user.

* Add shift management tests and enhance user validations

New feature tests have been introduced for verifying shift management operations including setting, starting and error-checking for shift locations. Also enhanced user validation logic in the Shift component to ensure an authenticated user and adjusted the Dashboard test to reflect this.

* Refactor shift handling and update tests

The shift handling process has been significantly refactored in this commit. Changes include the removal of redundant attribute validations in the Shift class, and moving validation rules to the functions where they are required. Additionally, the user authentication check has been moved to a private function. The Dashboard test has been updated to simulate a logged-in user, reflecting the changes in user validation.