Add spinner on login and prevent back history
Implemented a loading spinner that displays during the login process to enhance user experience. Also added a middleware to prevent browser back history for increased security, and adjusted the idle timeout values and modal display for improved interface responsiveness. A new middleware, PreventBackHistoryMiddleware, was created to manage browser cache settings and to prevent going back to the previous page.