Add Journal page and link to Navigation
A new Livewire component "Journal" is added, and the corresponding view 'journal.blade.php' is created. The 'journal' view displays an overview of all incidents. Furthermore, 'Journal' link is added to the navigation, enabling users to navigate to the Journal page. This link is only shown when the user is on duty. The 'journal' route is also added and authenticated.