Enhancement (#63)
* Add no-shift view for unauthorized access

The commit introduces a conditional rendering in the Dashboard class. It checks if a user has the 'work' authorization; if the user does not, the system will display the 'no-shift' view. This serves as a check and balance to prevent unauthorized shifts.

* Update database migration comment and remove team table

The comment relating to the 'role_user' table creation has been adjusted to clearly state that it is for associating roles to users only. In the same vein, the statement for dropping the 'teams' table during the migration rollback has been removed as it no longer fits the table structure.

* Add 'location' column to locations table

A 'location' column has been added to the locations table in the database migration file. This will allow each location to store more detailed geographical information, enhancing location tracking and management.