Split 'fullname' into 'firstname' and 'lastname' (#66)
* Split 'fullname' into 'firstname' and 'lastname'

In the users table migration, the 'fullname' column has been split into two separate columns - 'firstname' and 'lastname'. This change allows for a more precise identification and sorting of user records.

* Update UserFactory data structure

The UserFactory's definition method has been modified to separate first and last names. The fake name is now split into two parts, and these are used respectively for the 'firstname' and 'lastname' fields. The 'username' field is also assembled using the first and last names.