Refactor user shift and authorization states
The commit updates how user's shift and authorization states are handled. A new property, canWork, is added to track authorization for work at the start and end of the shift. Also, the usage of checkUserAuthorization() is replaced with checkUserShift() for better consistency. Some methods in User model are renamed and updated to offer clearer functionality, such as fetching shift data and creating time tracker.