Update ban until date when adding incident (#84)
Updated the Livewire AddIncident component to fix a bug where the 'ban until' date was not being validated against the existing 'ban until' date. Also added 'ban until' as a target in the wire loading attribute in the Laravel Blade file. This ensures that the 'ban until' date is updated correctly, and respects any existing 'ban until' date that the user might have, enhancing the incident adding functionality within the application.