Implemented Sqids for Journal entries (#88)
Integrated the Sqids package into the application, which allows the generation of unique short ID for Journal entries. This is represented as 'sqid' and has been added to the Journal Model. A new library, 'guava/sqids-for-laravel' has been added to composer.json. The Journal view has been updated to display the 'sqid' value.